Support for dht11, many fiddles

+++ ticks.h
@@ -1,5 +1,7 @@
+// "Millisecond" tick counter
+
 // (re)start the tick counter with specified initial value
-void start_ticks(unsigned long initial);
+void start_ticks(uint32_t initial);
 
 // stop the tick counter
 void stop_ticks(void);
